preprocessing organization
