+++ clinicadl/dataset/config/__init__.py
@@ -1,13 +1,1 @@
 from .file_type import FileType
-from .preprocessing import (
-    PreprocessingConfig,
-    PreprocessingCustom,
-    PreprocessingFlair,
-    PreprocessingPET,
-    PreprocessingT1,
-    PreprocessingT2,
-)
-from .utils import (
-    get_extraction,
-    get_preprocessing,
-)
