+++ clinicadl/dataset/preprocessing/__init__.py
@@ -1,0 +1,7 @@
+from .base import BasePreprocessing
+from .custom import PreprocessingCustom
+from .dti import PreprocessingDTI
+from .flair import PreprocessingFlair
+from .pet import PreprocessingPET
+from .t1 import PreprocessingT1
+from .t2 import PreprocessingT2
